update crud to use pg
