--- STYLE-GUIDE.md
@@ -1,6 +1,0 @@
-## Style Guide
-
-Refer to the [AirBnb Style Guide](https://github.com/airbnb/javascript).
-
-    // "babel-loader": "^7.1.5",
-    // "babel-preset-env": "^1.7.0",
