+++ server/server.js
@@ -1,0 +1,5 @@
+const app = require('./app.js');
+
+const PORT = process.env.PORT || 3002;
+
+app.listen(PORT, console.log(`App listening on port ${PORT}!`));
